Rename module dashboard/settings to dashboard/sites
