+++ templates/dashboard/base.html
@@ -64,7 +64,7 @@
                 <li class="{% block menu_product_classes_class %}{% endblock %}"><a href="{% url 'dashboard:product-class-list' %}">{% trans "Product types" %}</a></li>
                 <li class="{% block menu_attributes_class %}{% endblock %}"><a href="{% url 'dashboard:product-attributes' %}">{% trans "Attributes" %}</a></li>
                 <li class="{% block menu_delivery_class %}{% endblock %}"><a href="{% url 'dashboard:shipping-methods' %}">{% trans "Shipping methods" %}</a></li>
-                <li class="{% block menu_settings_class %}{% endblock %}"><a href="{% url 'dashboard:setting-index' %}">{% trans "Settings" %}</a></li>
+                <li class="{% block menu_settings_class %}{% endblock %}"><a href="{% url 'dashboard:site-index' %}">{% trans "Site settings" %}</a></li>
               </ul>
             </li>
           </ul>
