+++ templates/dashboard/sites/detail.html
