+++ templates/dashboard/sites/index.html
@@ -16,7 +16,7 @@
 {% block header_extra %}
     <div class="fixed-action-btn">
         <a class="btn-fab btn-floating btn-large waves-effect waves-light teal" title="{% trans "Add setting" %}"
-            href="{% url 'dashboard:setting-create' %}">
+            href="{% url 'dashboard:site-create' %}">
             <i class="material-icons">add</i>
         </a>
     </div>
